+++ daemon/volumes_unix.go
@@ -359,7 +359,7 @@ func (daemon *Daemon) registerMountPoints(container *Container, hostConfig *runc
 		if err := label.Relabel(bind.Source, container.MountLabel, shared); err != nil {
 			return err
 		}
-		if err := volume.FixUidGid(container.Config.User, bind.Source, bind.Mode); err != nil {
+		if err := volume.FixUIDGID(container.Config.User, bind.Source, bind.Mode); err != nil {
 			return err
 		}
 		binds[bind.Destination] = true
